requirements.txt
pydantic
typing
redis
dotenv
google.generativeai
